+++ .storybook/preview.ts
@@ -1,14 +1,10 @@
 import type { Preview } from '@storybook/web-components';
 import { addons } from '@storybook/preview-api';
-import { setCustomElementsManifest } from '@storybook/web-components';
 import { light, dark } from './forge-theme';
-import cem from '../dist/cem/custom-elements.json';
 
 import '@tylertech/forge/theme/forge-theme.scss';
 import '@tylertech/forge/forge-tokens.scss';
 import './preview-global.scss';
-
-setCustomElementsManifest(cem);
 
 const toggleDarkTheme = value => document.body.classList.toggle('forge-storybook-dark', value);
 const channel = addons.getChannel();
@@ -18,11 +14,10 @@ channel.off('DARK_MODE', isDark => toggleDarkTheme(isDark));
 const preview: Preview = {
   parameters: {
     docs: {
-      source: { format: 'html' },
+      source: { format: 'html', dark: true },
       toc: {
         contentsSelector: '.sbdocs-content',
-        headingSelector: 'h2,h3,h4',
-        ignoreSelector: '#primary'
+        headingSelector: 'h2,h3,h4'
       },
     },
     darkMode: { light, dark },
@@ -36,7 +31,9 @@ const preview: Preview = {
             'Usage',
             'Typography',
             'Theming',
+            'Accessibility',
             'Customization',
+            'Forms'
           ],
           'Frameworks',
           [